Changed try\except to Django shortcut

+++ mysite/publications/views.py
@@ -1,9 +1,8 @@
 from django.shortcuts import render, get_object_or_404
-from django.http import HttpResponse, Http404, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect
 from django.urls import reverse
-# Create your views here.
-from .models import Publication, Comment
 
+from .models import Publication, Comment
 
 def index(request):
     template_name = "publications/index.html"
@@ -15,28 +14,27 @@ def index(request):
 
 def detail(request, pk):
     # Try to get publication with pk or 404
-    try:
-        publication = Publication.objects.get(pk=pk)
-    except Publication.DoesNotExist:
-        raise Http404(f"The publication with {pk} as id does not exist.")
+    publication = get_object_or_404(Publication, pk=pk)
     # Pass the context to publication/detail.html
     publication = Publication.objects.get(pk=pk)
     text = publication.text
     title = publication.title
     pub_date = publication.pub_date
+    comments = publication.comment_set.all()
+    comment_num = len(comments)
     context = {
         "text": text,
         "title": title,
         "pub_date": pub_date,
-        "publication": publication
+        "publication": publication,
+        "comment_num": comment_num
     }
     return render(request, "publications/detail.html", context=context)
 
 def comment(request, pk):
-    try:
-        publication = Publication.objects.get(pk=pk)
-        comments = publication.comment_set.all()
-    except:
+    publication = get_object_or_404(Publication, pk=pk)
+    comments = publication.comment_set.all()
+    if not comments:
         return HttpResponse("No comments there yet")
     context = {
         "comments": comments,
@@ -45,11 +43,7 @@ def comment(request, pk):
     return render(request, "publications/comments.html", context=context)
 
 def add_comment(request, pk):
-    try:
-        publication = Publication.objects.get(pk=pk)
-    except Publication.DoesNotExist:
-        return HttpResponse("Publication doesn't exist")
-
+    publication = get_object_or_404(Publication, pk = pk)
     text = request.POST["text"]    
     new_comment = publication.comment_set.create(text=text)
     new_comment.save()
@@ -66,10 +60,7 @@ def add_publication(request):
     return HttpResponseRedirect(reverse("publications:index"))
 
 def del_comment(request, comment_id, pk):
-    try:
-        comment_to_delete = Comment.objects.get(id=comment_id)
-    except Comment.DoesNotExist:
-        return HttpResponse("Comment doesn't exist")
-    
-    comment_to_delete.delete()
+    comment_to_delete = get_object_or_404(Comment, pk=comment_id)
+    if comment_to_delete:
+        comment_to_delete.delete()
     return HttpResponseRedirect(reverse("publications:comment", args=(pk,)))
